Created a new full backup function and completely rewrite the handleRestore function to be "smart".
#### A. Rename handleBackup and Create handleFullBackup
Renamed the existing handleBackup to be more specific and then add the new full backup logic.
**Important:** Added a `type` property to the `backupData` object in both functions. This is the key to how the smart restore function will identify the file type.

#### B. The New "Smart" `handleRestore` Function
This is the core of the new feature. Replaced the entire existing `handleRestore` function with this new, more intelligent version.

Key Point: When importing the territory, we delete the id from the house and visit objects before adding them to the store. This is crucial because it forces IndexedDB to assign a new, unique ID within the local database, preventing conflicts.

#### C. Update Event Listeners in setupEventListeners
Finally, we need to connect our new and renamed functions to the correct buttons.

The backup and restore system is now significantly more powerful and flexible, perfectly supporting a more collaborative workflow.